[usdMaya, pxrUsdMayaGL] add shapeAdapters at end of Sync calls
...this ensures that the adapters are added to the batchRenderer, even
when we are using MAYA_VP2_USE_VP1_SELECTION: formerly, because the VP1
getDrawRequests call was never made, the VP1 adapter was never added to
the batchRenderer.  This would meant that when we do a rendererPlugin,
we would not clear the delete VP1 adapter's _delegate before deleting the
renderIndex.  Then, when the VP1 adapter was used by the select code, it
would notice it's delegate was out of date, and try to grab a new one...
...but would in the process delete it's old delegate, and the old
delegate's destructor would access the (now deleted) renderIndex